Implemented authorities for delete/change Hotel Implemented password forgotten and changing added actuator

+++ src/main/java/com/binarystudio/academy/springsecurity/domain/hotel/HotelService.java
@@ -1,6 +1,13 @@
 package com.binarystudio.academy.springsecurity.domain.hotel;
 
 import com.binarystudio.academy.springsecurity.domain.hotel.model.Hotel;
+import com.binarystudio.academy.springsecurity.domain.user.UserService;
+import com.binarystudio.academy.springsecurity.domain.user.model.User;
+import com.binarystudio.academy.springsecurity.domain.user.model.UserRole;
+import com.binarystudio.academy.springsecurity.security.auth.AuthoritiesException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,16 +17,19 @@ import java.util.UUID;
 @Service
 public class HotelService {
 	private final HotelRepository hotelRepository;
+	@Autowired
+	UserService userService;
 
 	public HotelService(HotelRepository hotelRepository) {
 		this.hotelRepository = hotelRepository;
 	}
 
 	public void delete(UUID hotelId) {
-		// 4. todo: only the owner of the hotel or admin may delete the hotel
-		boolean wasDeleted = hotelRepository.delete(hotelId);
-		if (!wasDeleted) {
-			throw new NoSuchElementException();
+		if (userHasPermission(hotelId)) {
+			boolean wasDeleted = hotelRepository.delete(hotelId);
+			if (!wasDeleted) {
+				throw new NoSuchElementException();
+			}
 		}
 	}
 
@@ -29,9 +39,11 @@ public class HotelService {
 
 
 	public Hotel update(Hotel hotel) {
-		// 4. todo: only the owner of the hotel or admin may update the hotel
-		getById(hotel.getId());
-		return hotelRepository.save(hotel);
+		if (userHasPermission(hotel.getId())) {
+			getById(hotel.getId());
+			return hotelRepository.save(hotel);
+		}
+		return null;
 	}
 
 	public Hotel create(Hotel hotel) {
@@ -40,5 +52,13 @@ public class HotelService {
 
 	public Hotel getById(UUID hotelId) {
 		return hotelRepository.getById(hotelId).orElseThrow();
+	}
+
+	private boolean userHasPermission(UUID hotelId){
+		User user = userService.getCurrentUser();
+		if (!user.getAuthorities().contains(UserRole.ADMIN) && !hotelRepository.userIsHotelOwner(user, hotelId)){
+			throw new AuthoritiesException();
+		}
+		return true;
 	}
 }
